LoRa/requirements.txt
meshcore
python-dotenv